\name{as.rlpspec}
\alias{as.rlpspec}
\title{Coerce an object into an 'rlpspec' object}
\usage{
  as.rlpspec(X)
}
\arguments{
  \item{X}{An object to coerce}
}
\description{
  Coerce an object into an 'rlpspec' object
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\seealso{
  \code{\link{is.rlpspec}}, \code{\link{rlpspec}}
}

