% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_coh_phase.R
\name{coherence}
\alias{coherence}
\title{coherence}
\usage{
coherence(pgram)
}
\arguments{
\item{pgram}{\code{numeric array} must be multivariate}
}
\value{
list of coherency
}
\description{
calculate coherency from the spectra and cross-spectra
}
