% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ctap_loess}
\alias{ctap_loess}
\alias{ctap_loess.tapers}
\title{Taper constraints using loess smoothing}
\usage{
ctap_loess(tapvec, ...)

\method{ctap_loess}{tapers}(tapvec, tapseq = NULL, loess.span = 0.3,
  loess.degree = 1, verbose = TRUE, ...)
}
\arguments{
\item{tapvec}{integer; the number of tapers at each frequency (can be a vector)}

\item{...}{additional arguments}

\item{tapseq}{vector; positions to evaluate derivatives (unused here, but necessary for smoother methods)}

\item{loess.span}{scalar; the span used in \code{loess}}

\item{loess.degree}{scalar; the polynomial degree}

\item{verbose}{logical; should warnings and messages be given?}
}
\description{
Taper constraints using loess smoothing
}
\seealso{
\code{\link{constrain_tapers}}, \code{\link{ctap_simple_rcpp}}
}
\keyword{tapers}
\keyword{tapers-constraints}

