% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{det_vector}
\alias{det_vector}
\title{det_vector}
\usage{
det_vector(x)
}
\arguments{
\item{x}{\code{numeric array} values to evaluate}
}
\value{
vector of determinants
}
\description{
Determinant for an array
}
