\name{is.taper}
\alias{is.taper}
\title{Reports whether x is a 'taper' object}
\usage{
  is.taper(x)
}
\arguments{
  \item{x}{An object to test}
}
\description{
  Reports whether x is a 'taper' object
}
\examples{
is.taper(1:10)
is.taper(as.taper(1:10))
}
\seealso{
  \code{\link{as.taper}}, \code{\link{taper}}
}

