% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parabolic_weights_field}
\alias{parabolic_weights_field}
\title{parabolic_weights_field}
\usage{
parabolic_weights_field(ntap)
}
\arguments{
\item{ntap}{the maximum number of tapers}
}
\description{
parabolic_weights_field
}
