% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/func_psdcore.R
\name{pgram_compare}
\alias{pgram_compare}
\alias{pgram_compare.amt}
\title{Plot an sine multitaper spectrum against a cosine-tapered periodogram}
\usage{
pgram_compare(x, y, ...)

\method{pgram_compare}{amt}(x, y, log.freq = TRUE, db.spec = TRUE,
  taper = 0.2, ...)
}
\arguments{
\item{x}{numeric; the frequencies to plot; alternatively, a single \code{\link{psdcore}} object}

\item{y}{a single \code{\link{psdcore}} object; optional if \code{x} is an appropriate set of frequencies}

\item{...}{additional parameters (currently unused)}

\item{log.freq}{logical; should frequencies be transformed with \code{\link{log10}}?
Note that if \code{x} is a set of frequencies, the values should not already be transformed.}

\item{taper}{numeric; specifies the proportion of data to taper for the cosine periodogram.}

\item{dp.spec}{logical; should the spectrum estimates be converted to decibels with \code{\link{dB}}?}
}
\value{
A list with the cosine-tapered estimates and the adaptive estimates, invisibly.
}
\description{
Plot the results of \code{\link{psdcore}} against the results of
\code{\link[stats]{spec.pgram}}
}

