% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_coherence_phase.R
\name{phase}
\alias{phase}
\title{phase}
\usage{
phase(pgram)
}
\arguments{
\item{pgram}{\code{numeric array} must be multivariate}
}
\value{
list of phase. For multivariate time series a matrix containing the cross-spectrum phase between different series. The format is the same as coh.
}
\description{
Calculate phase from the spectra and cross-spectra. This method is the same as
used in \code{spec.pgram}.
}
