% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resample_fft_rcpp}
\alias{resample_fft_rcpp}
\title{Resample an fft using varying tapers}
\usage{
resample_fft_rcpp(fftz, tapers, verbose = TRUE, dbl = TRUE,
  tapcap = 1000L)
}
\arguments{
\item{fftz}{complex; a vector representing the dual-length \code{\link[fftw]{FFT}}; see also \code{dbl}.}

\item{tapers}{integer; a vector of tapers}

\item{verbose}{logical; should messages be given?}

\item{dbl}{logical; should the code assume \code{fftz} is dual-length or singl-length?}

\item{tapcap}{integer; the maximum number of tapers which can be applied; note that the length is
automatically limited by the length of the series.}
}
\description{
Resample an fft using varying tapers
}

