% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{riedsid_rcpp}
\alias{riedsid_rcpp}
\title{replaces time consuming portion of riedsid2}
\usage{
riedsid_rcpp(PSD, ntaper)
}
\arguments{
\item{PSD}{vector or class \code{'amt'} or \code{'spec'}; the spectral values used to optimize taper numbers}

\item{ntaper}{scalar or vector; number of tapers to apply optimization}
}
\value{
kopt vector
}
\description{
replaces time consuming portion of riedsid2
}
