\docType{methods}
\name{rlpspec-methods}
\alias{lines}
\alias{plot}
\alias{print}
\alias{rlpspec}
\alias{rlpspec-methods}
\alias{summary}
\title{Generic methods for 'rlpspec' objects}
\usage{
  print.summary.rlpspec(x, ...)

  print.rlpspec(psd, ...)
}
\description{
  Generic methods for 'rlpspec' objects
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\seealso{
  \code{\link{is.rlpspec}}, \code{\link{as.rlpspec}}
}
\keyword{generic}
\keyword{methods}

