\name{newRlpspec}
\alias{newRlpspec}
\alias{rlpspec}
\alias{rlpspec-class,}
\title{The 'rlpspec' S4 class.}
\usage{
  newRlpspec(...)
}
\description{
  This class very nearly resembles the structure of a
  'spec' object; however, it has slots to store information
  about tapers, the adaptive procedure, and any ancillary
  processing (e.g. prewhitening, etc) which may have been
  done on the original series.
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}

