% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd-package.R
\docType{data}
\name{wipp30}
\alias{wipp30}
\title{A single line of Project MAGNET horizontal field intensity}
\format{
A matrix with 13413 rows following 4 variables.

\describe{
\item{\code{time}}{Time (hours)}
\item{\code{wl}}{Water levels (psi)}
\item{\code{baro}}{Barometric pressure (psi)}
\item{\code{et}}{Earth tide gravity potential (nanometers/second^2)}
}
}
\source{
BETCO page: \url{http://www.hydrology.uga.edu/rasmussen/betco/}
}
\description{
The Project MAGNET mission 
provided a wealth of airborne-magnetometer data
spanning the globe (Coleman, 1992).  
This dataset represents a single track of horizontal field
intensities (a very small subset of the full collection!).
}
\details{
\subsection{Raw and Clean Sets}{
There are non-real data points in raw MAGNET series; these are 
instrumental artefacts, and can severely affect
power spectral density (PSD) estimates.  
A clean series has been included
so that a comparison of PSDs may be made.

Some command like \code{subset(magnet, abs(mdiff) > 0)}
can be used to identify the rows where edits have been made.
}
}
\examples{
data(wipp30)
summary(wipp30)
}
\references{
Toll, N.J., Rasmussen, T.C., (2007),
 Removal of Barometric Pressure Effects and Earth Tides from Observed Water Levels.
 \emph{Ground Water}, \strong{45}, 101–105. \url{https://doi.org/10.1111/j.1745-6584.2006.00254.x}
}
\seealso{
\code{\link{pspectrum}}
}
\keyword{datasets}
