% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_coherence_phase.R
\name{coherence}
\alias{coherence}
\title{coherence}
\usage{
coherence(pgram)
}
\arguments{
\item{pgram}{\code{numeric array} must be multivariate}
}
\value{
list of coherence. For multivariate time series, a 
matrix containing the squared coherency between different series. 
Column i + (j - 1) * (j - 2)/2 contains the squared coherency 
between columns i and j of x, where i < j.
}
\description{
Calculate coherence from the spectra and cross-spectra. This method is the same as
used in \code{spec.pgram}.
}
