% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_tapers.R
\name{ctap_loess}
\alias{ctap_loess}
\alias{ctap_loess.tapers}
\alias{ctap_loess.default}
\title{Taper constraints using loess smoothing}
\usage{
ctap_loess(tapvec, ...)

\method{ctap_loess}{tapers}(tapvec, ...)

\method{ctap_loess}{default}(
  tapvec,
  tapseq = NULL,
  loess.span = 0.3,
  loess.degree = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{tapvec}{integer or \code{'tapers'} object; the number of tapers at each frequency}

\item{...}{optional arguments}

\item{tapseq}{numeric; positions or frequencies -- necessary for smoother methods}

\item{loess.span}{scalar; the span used in \code{loess}}

\item{loess.degree}{scalar; the polynomial degree}

\item{verbose}{logical; should warnings and messages be given?}
}
\description{
Taper constraints using loess smoothing
}
\seealso{
\code{\link{constrain_tapers}}, \code{\link{ctap_simple}}
}
