% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_coherence_phase.R
\name{phase}
\alias{phase}
\title{phase}
\usage{
phase(pgram)
}
\arguments{
\item{pgram}{\code{numeric array} must be multivariate}
}
\value{
list of phase
}
\description{
calculate phase from the spectra and cross-spectra
}
