\name{prewhiten}
\alias{prewhiten}
\alias{prewhiten.default}
\alias{prewhiten.ts}
\title{Prewhiten a series.}
\usage{
  prewhiten(tser, AR.max = 0L, detrend = TRUE,
    demean = TRUE, plot = TRUE, verbose = TRUE,
    x.fsamp = 1, x.start = c(1, 1), ...)

  \method{prewhiten}{default} (tser, AR.max = 0L,
    detrend = TRUE, demean = TRUE, plot = TRUE,
    verbose = TRUE, x.fsamp = 1, x.start = c(1, 1), ...)

  \method{prewhiten}{ts} (tser, AR.max = 0L,
    detrend = TRUE, demean = TRUE, plot = TRUE,
    verbose = TRUE, x.fsamp = NA, x.start = NA, ...)
}
\arguments{
  \item{tser}{vector; An object to prewhiten.}

  \item{AR.max}{numeric; the maximum AR order to fit.}

  \item{detrend}{logical; Should a trend (and mean) be
  removed?}

  \item{demean}{logical; Should a mean value be removed?}

  \item{plot}{logical; Should the results be plotted?}

  \item{verbose}{logical; Should messages be printed?}

  \item{x.fsamp}{sampling frequency (for non \code{ts}
  objects)}

  \item{x.start}{start time of observations (for non
  \code{ts} objects)}

  \item{...}{variables passed to \code{prewhiten.ts} (for
  non \code{ts} objects)}
}
\description{
  Remove (optionally) mean, trend, and Auto Regressive (AR)
  model from the original series.
}
\details{
  The R-S multitapers do not exhibit the remarkable
  spectral-leakage suppression properties of the Thomson
  prolate tapers, so that in spectra with large dynamic
  range, power bleeds from the strong peaks into
  neighboring frequency bands of low amplitude -- spectral
  leakage. Prewhitening can ameliorate the problem, at
  least for red spectra [see Chapter 9, Percival and Walden
  (1993)].

  This function has essentially two modes of operation
  (detailed below): \describe{ \item{\code{AR.max} <
  1}{Remove (optionally) a mean and/or linear trend.}
  \item{\code{AR.max} > 0}{Remove an autoregressive model}
  } In the second case, the time series is filtered in the
  time domain with a finite-impluse-response filter of
  \code{AR.max} terms. The filter is found by solving the
  Yule-Walker equations for which it is assumed the series
  was generated by an autoregressive process, up to order
  \code{AR.max}. \subsection{Mean and trend (\code{AR.max <
  1})}{ Power spectral density estimates can become badly
  biased (especially at lower frequencies) if a signal of
  the form \eqn{f(x) = A x + B} is not removed from the
  series. If \code{detrend=TRUE} a model of this form is
  removed over the entire series using a linear
  least-squares estimator; in this case a mean value is
  removed regardless of the logical state of \code{demean}.
  To remove \emph{only} a mean value, set
  \code{detrend=FALSE} and (obviously) \code{demean=TRUE}.
  } \subsection{Auto Regressive (AR) innovations
  (\code{AR.max > 0})}{ When an autoregressive model is
  removed from a non-stationary series, the residuals are
  known as 'innovations', and may be stationary (or
  very-nearly stationary). This function fits an AR model
  [order at least 1, but up to and including
  AR(\code{AR.max})] to the series by solving the
  Yule-Walker equations; however, AIC is used to estimate
  the highest significant order, which means that
  higher-order components may not necessarily be fit. The
  resulting innovations can be used to better estimate the
  stationary component of the original signal, and possibly
  in an interactive editing method.

  A quick way to determine whether this may be needed for
  the series is to run \code{acf} on it, and see if
  significant auto-correlations are found. }
}
\section{NA values}{
  \code{NA} values are allowed.  If present, the function
  in the package \code{zoo}, \code{na.locf}, which stands
  for "Last Observation Carried Forward", is used to impute
  them with real numbers.
}
\author{
  A.J. Barbour <andy.barbour@gmail.com> and Robert L.
  Parker
}
\seealso{
  \code{\link{psdcore}}, \code{\link{pspectrum}}
}
\keyword{S3methods}
\keyword{autoregressive-innovations}
\keyword{prewhiten}
\keyword{timeseries}

