% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ctap_simple}
\alias{rcpp_ctap_simple}
\title{c++ implementation of constrain filter}
\usage{
rcpp_ctap_simple(tapvec, maxslope = 1L)
}
\arguments{
\item{tapvec,maxslope}{see \code{\link{ctap_simple}}}
}
\description{
c++ implementation of constrain filter
}

