% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_ctap_simple}
\alias{rcpp_ctap_simple}
\title{c++ implementation of the RLP constraint filter}
\usage{
rcpp_ctap_simple(tapvec, maxslope = 1L)
}
\arguments{
\item{tapvec}{integer or \code{'tapers'} object; the number of tapers at each frequency}

\item{maxslope}{integer; constrain based on this maximum first difference}
}
\description{
c++ implementation of the RLP constraint filter
}
