\docType{rlpSpec}
\name{rlpSpec}
\alias{rlpSpec}
\alias{spec.rlp}
\title{what this does}
\description{
  This package provides tools to
}
\details{
  The main function to be used is \code{\link{some_func}}

  There are also two helper functions included: \describe{
  \item{\code{\link{some_other_func}}}{ to do something.} }
}
\section{Scientific background}{
  A bunch of stuff, and inline equation \eqn{r}, and a
  newline equation: \deqn{ \frac{\partial^2 s}{\partial
  r^2}= 0 } and some more.

  And more.
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  Riedel, K. S., & Sidorenko, A. (1995), Minimum bias
  multiple taper spectral estimation, \emph{Signal
  Processing, IEEE Transactions on}, \strong{43}(1),
  188-195.

  Riedel, K. S. (1996), Adaptive smoothing of the
  log-spectrum with multiple tapering, \emph{Signal
  Processing, IEEE Transactions on}, \strong{44}(7),
  1794-1800.

  \url{http://igppweb.ucsd.edu/%7Eparker/Software/#PSD}
}
\seealso{
  \code{\link{some_function}},
  \code{\link{some_other_func}}
}

