\name{taper-methods}
\alias{plot}
\alias{print}
\alias{summary}
\alias{taper}
\alias{taper-methods}
\title{Generic methods for 'taper' objects}
\usage{
  print.taper(x)

  summary.taper(x)

  print.summary.taper(x)

  plot.taper(x, color.pal = c("Blues", "Spectral"), ...)
}
\description{
  Generic methods for 'taper' objects
}
\seealso{
  \code{\link{is.taper}}, \code{\link{as.taper}}
}
\keyword{generic}
\keyword{methods}

